# document::createFromString optimizations