--- a/include/litehtml/background.h
+++ b/include/litehtml/background.h
@@ -2,7 +2,6 @@
 #define LH_BACKGROUND_H
 
 #include "types.h"
-#include "attributes.h"
 #include "css_length.h"
 #include "css_position.h"
 #include "web_color.h"