--- a/include/litehtml/borders.h
+++ b/include/litehtml/borders.h
@@ -33,7 +33,7 @@ namespace litehtml
 			return *this;
 		}
 
-		string to_string();
+		string to_string() const;
 	};
 
 	struct border
@@ -250,7 +250,7 @@ namespace litehtml
 			radius	= val.radius;
 			return *this;
 		}
-		string to_string()
+		string to_string() const
 		{
 			return	"left: " + left.to_string() +
 					", top: " + top.to_string() +