--- a/include/litehtml/css_margins.h
+++ b/include/litehtml/css_margins.h
@@ -31,7 +31,7 @@ namespace litehtml
 			return *this;
 		}
 
-		string to_string()
+		string to_string() const
 		{
 			return	"left: " + left.to_string() +
 					", right: " + right.to_string() +