--- a/include/litehtml/css_selector.h
+++ b/include/litehtml/css_selector.h
@@ -190,15 +190,13 @@ namespace litehtml
 		css_element_selector	m_right;
 		css_selector::ptr		m_left;
 		css_combinator			m_combinator;
-		string					m_style;
+		style::ptr				m_style;
 		int						m_order;
 		media_query_list::ptr	m_media_query;
-		string					m_baseurl;
 	public:
-		explicit css_selector(const media_query_list::ptr& media, const string& baseurl)
+		explicit css_selector(const media_query_list::ptr& media = nullptr)
 		{
 			m_media_query	= media;
-			m_baseurl		= baseurl;
 			m_combinator	= combinator_descendant;
 			m_order			= 0;
 		}