--- a/include/litehtml/string_id.h
+++ b/include/litehtml/string_id.h
@@ -168,8 +168,12 @@ STRING_ID(
 	_background_origin_,
 	_background_clip_,
 	_background_attachment_,
-	_background_position_,
 	_background_size_,
+	__litehtml_background_width_,
+	__litehtml_background_height_,
+	_background_position_,
+	_background_position_x_,
+	_background_position_y_,
 
 	_border_,
 	_border_width_,