--- a/include/litehtml/style.h
+++ b/include/litehtml/style.h
@@ -3,32 +3,66 @@
 
 namespace litehtml
 {
+	enum property_type
+	{
+		prop_type_invalid, // indicates "not found" condition in style::get_property
+		prop_type_inherit, // "inherit" was specified as the value of this property
+
+		prop_type_enum_item,
+		prop_type_length,
+		prop_type_number,
+		prop_type_color,
+		prop_type_string,
+
+		prop_type_var, // also string, but needs further parsing because of var()
+	};
+
 	class property_value
 	{
 	public:
-		string	m_value;
+		property_type	m_type;
+
+		int 			m_enum_item;
+		css_length		m_length;
+		float			m_number;
+		web_color		m_color;
+		string			m_string;
+
 		bool			m_important;
 
 		property_value()
 		{
-			m_important = false;
+			m_type = prop_type_invalid;
 		}
-		property_value(const char* val, bool imp)
+		property_value(const string& str, bool important)
 		{
-			m_important = imp;
-			m_value		= val;
+			m_type = prop_type_string;
+			m_string = str;
+			m_important = important;
 		}
-		property_value(const property_value& val)
+		property_value(const css_length& length, bool important)
 		{
-			m_value		= val.m_value;
-			m_important	= val.m_important;
+			m_type = prop_type_length;
+			m_length = length;
+			m_important = important;
 		}
-
-		property_value& operator=(const property_value& val)
+		property_value(float number, bool important)
 		{
-			m_value		= val.m_value;
-			m_important	= val.m_important;
-			return *this;
+			m_type = prop_type_number;
+			m_number = number;
+			m_important = important;
+		}
+		property_value(int enum_item, bool important)
+		{
+			m_type = prop_type_enum_item;
+			m_enum_item = enum_item;
+			m_important = important;
+		}
+		property_value(web_color color, bool important)
+		{
+			m_type = prop_type_color;
+			m_color = color;
+			m_important = important;
 		}
 	};
 
@@ -41,32 +75,16 @@ namespace litehtml
 		typedef std::vector<style::ptr>		vector;
 	private:
 		props_map							m_properties;
+		static std::map<string_id, string>	m_valid_values;
 	public:
-		style() = default;
-		style(const style& val);
-
-		style& operator=(const style& val)
+		void add(const string& txt, const string& baseurl = "", document_container* container = nullptr)
 		{
-			m_properties = val.m_properties;
-			return *this;
+			parse(txt, baseurl, container);
 		}
 
-		void add(const char* txt, const char* baseurl = "", const element* el = nullptr)
-		{
-			parse(txt, baseurl, el);
-		}
+		void add_property(string_id name, const string& val, const string& baseurl = "", bool important = false, document_container* container = nullptr);
 
-		void add_property(string_id name, const char* val, const char* baseurl = "", bool important = false, const element* el = nullptr);
-
-		const char* get_property(string_id name) const
-		{
-			auto f = m_properties.find(name);
-			if(f != m_properties.end())
-			{
-				return f->second.m_value.c_str();
-			}
-			return nullptr;
-		}
+		const property_value& get_property(string_id name) const;
 
 		void combine(const style& src);
 		void clear()
@@ -75,12 +93,19 @@ namespace litehtml
 		}
 
 	private:
-		void parse_property(const string& txt, const char* baseurl, const element* el);
-		void parse(const char* txt, const char* baseurl, const element* el);
-		void parse_short_background(const string& val, const char* baseurl, bool important);
-		void parse_short_font(const string& val, bool important);
+		void parse_property(const string& txt, const string& baseurl, document_container* container);
+		void parse(const string& txt, const string& baseurl, document_container* container);
+		void parse_background(const string& val, const string& baseurl, bool important, document_container* container);
+		void parse_background_position(const string& val, bool important);
+		void parse_background_size(const string& val, bool important);
+		void parse_font(const string& val, bool important);
+		void parse_flex(const string& val, bool important);
+		static css_length parse_border_width(const string& str);
+		static void parse_two_lengths(const string& str, css_length len[2]);
+		static int parse_four_lengths(const string& str, css_length len[4]);
 		static void subst_vars(string& str, const element* el);
-		void add_parsed_property(string_id name, const string& val, bool important);
+
+		void add_parsed_property(string_id name, const property_value& propval);
 		void remove_property(string_id name, bool important);
 	};
 }