Result on Obama wiki:
```
document::createFromString: 1730 ms -> 1050 ms
```

Most of the speed-up (450 ms) is from undoing commit 5a00888b.

Another 240 ms is from doing almost all the css property parsing during stylesheet parsing.
The only parsing left for later is, by necessity, HTML `style` attribute and CSS properties containing `var`.
`parse_styles` doesn't do much parsing now, it was renamed to `compute_styles`.
All the css properties are supposed to be accessed through css() now, there is no html_tag::get_cursor or html_tag::get_color.
Instead of get_style_property there are several typed alternatives: get_enum_property, get_length_property, etc., but these functions are supposed to be called only from css_properties::compute, all the access from outside is through element::css().

Inherited properties are copied from the parent's m_css. I use offsets instead of pointers to members because I need to poke inside 
`css_properties` subobjects, like `m_css_margins.left`.

I use `css_length` for `m_z_index` because it can be `auto`.

Removed checks `if(m_display == display_list_item)`,
`if (m_border_collapse == border_collapse_separate)`,
`if (m_el_position != element_position_static)` from css_properties::compute
because descendants may need these properties (this is especially true for lists).


Other changes: removed copy constructors and copy assignment operators from property_value, style, and css_length because implicitly-defined ones do what is needed.